backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
face-recognition==1.3.0
dlib==19.24.2
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
pytesseract
easyocr
winsdk
psycopg2-binary==2.9.9
